// src/generated/api/models/IntroResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 모든 API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0.4
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface IntroResponse
 */
export interface IntroResponse {
    /**
     * 
     * @type {number}
     * @memberof IntroResponse
     */
    introId?: number;
    /**
     * 
     * @type {number}
     * @memberof IntroResponse
     */
    userId?: number;
    /**
     * 
     * @type {string}
     * @memberof IntroResponse
     */
    nickname?: string;
    /**
     * 
     * @type {string}
     * @memberof IntroResponse
     */
    profileImageUrl?: string;
    /**
     * 
     * @type {string}
     * @memberof IntroResponse
     */
    content?: string;
    /**
     * 
     * @type {Date}
     * @memberof IntroResponse
     */
    createdAt?: Date;
}

/**
 * Check if a given object implements the IntroResponse interface.
 */
export function instanceOfIntroResponse(value: object): value is IntroResponse {
    return true;
}

export function IntroResponseFromJSON(json: any): IntroResponse {
    return IntroResponseFromJSONTyped(json, false);
}

export function IntroResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): IntroResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'introId': json['introId'] == null ? undefined : json['introId'],
        'userId': json['userId'] == null ? undefined : json['userId'],
        'nickname': json['nickname'] == null ? undefined : json['nickname'],
        'profileImageUrl': json['profileImageUrl'] == null ? undefined : json['profileImageUrl'],
        'content': json['content'] == null ? undefined : json['content'],
        'createdAt': json['createdAt'] == null ? undefined : (new Date(json['createdAt'])),
    };
}

export function IntroResponseToJSON(json: any): IntroResponse {
    return IntroResponseToJSONTyped(json, false);
}

export function IntroResponseToJSONTyped(value?: IntroResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'introId': value['introId'],
        'userId': value['userId'],
        'nickname': value['nickname'],
        'profileImageUrl': value['profileImageUrl'],
        'content': value['content'],
        'createdAt': value['createdAt'] == null ? undefined : ((value['createdAt']).toISOString()),
    };
}